OpenEOClient

Documentation for OpenEOClient.

Tutorial

Processing data on a openEO server requires authentication.

using OpenEOClient
username = ENV["username"]
password = ENV["password"]
c = connect("earthengine.openeo.org", "v1.0", username, password)
step1 = c.load_collection(
    "COPERNICUS/S2", BoundingBox(west=16.06, south=48.06, east=16.65, north=48.35),
    ["2020-01-01", "2020-01-31"], ["B10"]
)
step2 = c.reduce_dimension(step1, Reducer("median"), "t", nothing)
step3 = c.save_result(step2, "GTIFF-ZIP", Dict())
path = c.compute_result(step3)

The data is downloaded in zipped GeoTiff format. It can be loaded into a local Julia session using Rasters.jl:

using ZipFile, Rasters, Plots, ArchGDAL

f = ZipFile.Reader(path).files[1]
write(open(f.name, "w"), read(f, String))

cube = Raster(f.name)

Plotting:

plot(cube)

OpenEOClient.AuthorizedConnection

OpenEOClient.AuthorizedConnection

OpenEOClient.ProcessGraph

OpenEOClient.ProcessGraph

OpenEOClient.compute_result

OpenEOClient.describe_collection

OpenEOClient.get_acces_token

OpenEOClient.list_collections

OpenEOClient.list_jobs

OpenEOClient.list_processes

OpenEOClient.AuthorizedConnection — Method

HTTP basic authentification

OpenEOClient.AuthorizedConnection — Method

OpenID Connect device flow + PKCE authentification

OpenEOClient.ProcessGraph — Type

Create a ProcessGraph to reduce dimesnions

OpenEOClient.ProcessGraph — Type

Create a ProcessGraph using a user-defined Function

This is useful e.g. to create reducers to combine values of different bands in a customized way.

OpenEOClient.compute_result — Function

Process and download data synchronously

OpenEOClient.describe_collection — Method

Lists all information about a specific collection specified by the identifier

OpenEOClient.get_acces_token — Method

OpenID Connect device flow + PKCE

OpenEOClient.list_collections — Method

Lists available collections with at least the required information.

OpenEOClient.list_jobs — Method

Lists all batch jobs submitted by a user.

OpenEOClient.list_processes — Method

Lists all predefined processes and returns detailed process descriptions, including parameters and return values.