OpenEOClient

Documentation for OpenEOClient.

OpenEOClient.describe_collection

OpenEOClient.list_collections

OpenEOClient.list_jobs

OpenEOClient.list_processes

OpenEOClient.save_result

OpenEOClient.describe_collection — Method

Lists all information about a specific collection specified by the identifier

OpenEOClient.list_collections — Method

Lists available collections with at least the required information.

OpenEOClient.list_jobs — Method

Lists all batch jobs submitted by a user.

OpenEOClient.list_processes — Method

Lists all predefined processes and returns detailed process descriptions, including parameters and return values.

OpenEOClient.save_result — Method

Process and download data synchronously